requirements/requirements.txt
uwsgi==2.0.18
flask-cors==3.0.8
Werkzeug==0.16.1
pymongo==3.10.1
flask==1.1.2
flask_script==2.0.6
flask_migrate==2.1.1
marshmallow==2.14.0
flask_sqlalchemy==2.3.2
flask_marshmallow==0.8.0
marshmallow-sqlalchemy==0.13.2
psycopg2-binary==2.8.4
pytz==2019.3
pandas==0.24.2
flask-restplus==0.13.0

git+ssh://git@github.com/eduardocurado/stock_models.git@master
git+ssh://git@github.com/eduardocurado/parser_broker.git@develop